+++ platforms/craft/craft/config/imagetransforms.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+
+    'pagebuilder' => [
+        ['width' => 1200],
+        ['width' => 1000],
+        ['width' => 800],
+        ['width' => 600, 'jpegQuality' => 80],
+        ['width' => 400, 'jpegQuality' => 80],
+    ],
+
+];